Remove TC mention from Security + Tweak phrasing
The semantics of the sentence pre-change suggest that both ACT and TC Desktop read game memory and networking packets.
I'm not sure how to word this to clarify that TC Desktop only reads game packets and does not read memory, so I just removed
it from the section altogether. We can add it back if we can rephrase this elegantly enough to convey the difference without
running on too long. I also changed the phrasing of the packets bit to clarify that Dalamud allows you to read all packets, and
not just incoming ones.